+++ packages/data-sdk/src/message-bus/listeners/addOverviewDeviceListener.ts
@@ -1,0 +1,17 @@
+import { EmbeddedAppMessage, IDevice } from "./EmbeddedAppMessage";
+import { sendAppMessage } from "../senders/sendAppMessage";
+
+export function addOverviewDeviceListener(
+  handler: (devices: IDevice[]) => void
+) {
+  sendAppMessage({ type: "request_devices" });
+  const listener = (event: MessageEvent<EmbeddedAppMessage>) => {
+    const msg = event.data;
+    if (msg.type === "overview_devices") {
+      handler(msg.data);
+    }
+  };
+
+  window.addEventListener("message", listener);
+  return () => window.removeEventListener("message", listener);
+}
