+++ packages/data-sdk/src/message-bus/senders/disableAnalyticsBottomBar.ts
@@ -1,0 +1,7 @@
+import { sendAppMessage } from "./sendAppMessage";
+
+export function disableAnalyticsBottomBar() {
+  sendAppMessage({
+    type: "hide_analytics_date_picker",
+  });
+}
