+++ packages/data-sdk/src/message-bus/senders/goToDevice.ts
@@ -1,0 +1,8 @@
+import { sendAppMessage } from "./sendAppMessage";
+
+export function goToDevice(deviceId: string) {
+  sendAppMessage({
+    type: "go_to_device",
+    deviceId,
+  });
+}
