+++ packages/data-sdk/src/message-bus/senders/goToTime.ts
@@ -1,0 +1,8 @@
+import { sendAppMessage } from "./sendAppMessage";
+
+export function goToTime(date: Date): void {
+  sendAppMessage({
+    type: "go_to_time",
+    time: date.getTime(),
+  });
+}
