+++ packages/data-sdk/src/message-bus/senders/showMessage.ts
@@ -1,0 +1,5 @@
+import { sendAppMessage } from "./sendAppMessage";
+
+export function showMessage(message: string) {
+  sendAppMessage({ type: "show_message", message });
+}
